delete is2cat and references
